navbar scss file created
