+++ src/Components/App/NavBar/Nav.js
@@ -1,0 +1,24 @@
+import { useContextProvider } from "../../../Providers/Provider";
+import { useNavProvider } from "../../../Providers/NavProvider";
+import SlideNav from "./SlideNav/SlideNav";
+import Header from "../../Job/Header.js"
+import hamburger from "../../../Assets/blue-hamburger-menu.png"
+import "./Nav.scss";
+
+export default function Nav() {
+  const { appHeader } = useContextProvider();
+  const { navbarClick } = useNavProvider()
+
+  return (
+    <nav className="navBar grid-center">
+      <img 
+      src={hamburger} alt="hamburger-menu"
+      className="navBar_burger" 
+      onClick={() => navbarClick()} />
+
+      <Header header={appHeader} />
+      
+      <SlideNav />
+    </nav>
+  );
+}
