+++ src/Components/App/NavBar/SlideNav/DarkModeSlider/DarkModeSlider.js
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import { useContextProvider } from "../../Providers/Provider";
-import { useNavProvider } from "../../Providers/NavProvider";
+import { useContextProvider } from "../../../../../Providers/Provider";
+import { useNavProvider } from "../../../../../Providers/NavProvider";
 import "./DarkModeSlider.scss";
 
 function DarkModeSlider() {
